archive/old_monolith/backend/requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.10.0
gunicorn==23.0.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pillow==12.0.0
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.5.0
vaderSentiment==3.3.2
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
detoxify
rank-bm25
sentence-transformers
faiss-cpu
